requirements/staging.txt
Django
django-cors-headers
django-extra-fields
django-filter
djangorestframework
pytz
sqlparse
Unipath
Pillow
django-cors-headers
django-quill-editor
